--- app/Providers/EventServiceProvider.php
@@ -1,45 +1,0 @@
-<?php
-
-namespace App\Providers;
-
-use App\Event;
-use App\Listeners\LogSentMessage;
-use App\Observers\EventObserver;
-use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Illuminate\Mail\Events\MessageSending;
-
-class EventServiceProvider extends ServiceProvider
-{
-    /**
-     * The event listener mappings for the application.
-     *
-     * @var array
-     */
-    protected $listen = [
-        'App\Events\Event' => ['App\Listeners\EventListener'],
-        \SocialiteProviders\Manager\SocialiteWasCalled::class => [
-            // add your listeners (aka providers) here
-            'SocialiteProviders\\Azure\\AzureExtendSocialite@handle',
-        ],
-
-        MessageSending::class => [LogSentMessage::class],
-    ];
-
-    /**
-     * Register any events for your application.
-     */
-    public function boot(): void
-    {
-        parent::boot();
-
-        Event::observe(EventObserver::class);
-    }
-
-    /**
-     * Determine if events and listeners should be automatically discovered.
-     */
-    public function shouldDiscoverEvents(): bool
-    {
-        return false;
-    }
-}
